requirements.txt
requests==2.31.0
asyncio~=3.4.3
python-dotenv~=1.0.0
discord==2.2.3